requirementsNew3.txt
asgiref==3.2.10
attrs==21.2.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cryptography==36.0.1
cymem==2.0.6
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.1
django-crispy-forms==1.13.0
django-extensions==3.1.3
django-import-export==2.5.0
django-inspectdb==0.2.1
django-model-utils==4.2.0
django-multiselectfield==0.1.12
django-object-tool==0.0.10
django-pagination==1.0.7
django-sass-processor==1.0.1
django-smtp-ssl==1.0
dnspython==1.16.0
docx2pdf==0.1.7
docx2txt==0.8
et-xmlfile==1.1.0
exception==0.1.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
Jinja2==3.0.3
joblib==1.0.1
jsonfield==3.1.0
jsonschema==4.3.2
langcodes==3.3.0
MarkupPy==1.14
MarkupSafe==2.0.1
murmurhash==1.0.6
nltk==3.6.6
numpy==1.21.5
odfpy==1.4.1
openpyxl==3.0.9
optionaldict==0.1.2
packaging==21.3
pandas==1.3.5
pathy==0.6.1
pdfminer.six==20211012
pdfplumber==0.6.0
phonenumbers==8.12.39
Pillow==8.4.0
preshed==3.0.6
pycparser==2.20
pycryptodome==3.12.0
pydantic==1.8.2
pyparsing==2.4.7
Pyphen==0.10.0
pyresparser==1.0.6
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2021.1
PyYAML==6.0
regex==2021.11.10
requests==2.25.1
resume-parser==0.8.2
six==1.16.0
smart-open==5.2.1
sortedcontainers==2.4.0
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.1
srsly==2.4.2
stemming==1.0.1
tablib==3.1.0
thinc==8.0.13
tika==1.24
tinycss2==1.1.0
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.5
Wand==0.6.7
wasabi==0.9.0
webencodings==0.5.1
wechat==0.4.16
wechat-django==0.3.3
wechatpy==1.8.3
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.12.0
